A modular framework for semi-automated supervised feature tracking
The core of this framework is the BrightObjectTracker which takes as arguments a loadFunction and a FindFeatures functiont and then goes frame by frame and loads in frames and extracts feature candidates. The user than goes in and approves the feature candidates are manually enters his or her own feature.

So far the software will load in some images and display the first 5 brightest 5pixel-radius regions.